pyproject.toml
[project]
name = "bloqade"
version = "0.1.0"
description = "The software development toolkit for neutral atom arrays."
readme = "README.md"
authors = [
    { name = "Roger-luo", email = "rluo@quera.com" },
    { name = "kaihsin", email="khwu@quera.com" },
    { name = "weinbe58", email="pweinberg@quera.com"},
    { name = "johnzl-777", email="jlong@quera.com"},
]
requires-python = ">=3.10"
dependencies = [
    "kirin-toolchain>=0.9.0",
]

[project.optional-dependencies]
qasm2 = [
    "lark>=1.2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/bloqade"]

[dependency-groups]
dev = [
    "black>=24.10.0",
    "coverage>=7.6.4",
    "ipython>=8.29.0",
    "isort>=5.13.2",
    "mypy>=1.13.0",
    "numpy>=2.1.3",
    "pre-commit>=4.0.1",
    "pyright>=1.1.388",
    "pytest>=8.3.3",
    "ruff>=0.7.3",
    "rust-just>=1.36.0",
    "tomlkit>=0.13.2",
]
doc = [
    "griffe-inherited-docstrings>=1.1.1",
    "mike>=2.1.3",
    "mkdocs>=1.6.1",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.1",
    "mkdocs-material>=9.5.44",
    "mkdocs-minify-plugin>=0.8.0",
    "mkdocstrings[python]>=0.27.0",
]

[tool.isort]
profile = "black"
combine_as_imports = true
multi_line_output = 3
length_sort = true
src_paths = ["src/kirin"]

[tool.black]
line-length = 88

[tool.ruff]
target-version = "py310"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.lint]
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.coverage.run]
include = ["src/kirin/*"]
